Add unit tests for shard creation
This required a special approach. Because the default behavior of PGXS-
based extensions is to use the installcheck target (which tests against
an existing PostgreSQL installation), we cannot assume anything about
the user's cluster. But we don't want to clobber their config, either.

So what to do? pg_regress has an option to use a launcher script rather
than calling psql directly. I've added one which:

  1. Moves the worker list configuration out of the way
  2. Replaces it with a known configuration
  3. Runs tests
  4. Moves the original configuration back into place

This is tedious, but works well to exercise all branches without any
knowledge of the user's cluster configuration.

Adding such a launcher allows us to test the shard creation logic,
which—while still modular—is the first end-to-end test thus far: we
actually build tables for shards and insert metadata to reflect this.