Add shard repair unit tests; fix bad strncmp use
This test builds a shard, puts some data in it, marks it as unhealthy,
and runs the repair command on its placement. We verify that the shard
has been rebuilt by comparing its OID with the OID of the pre-repair
relation.

A little hacky, but it's the best we can do without more scaffolding.