Misleading error for incompletely created tables
Before this fix, if you try to execute any query on tables which
are distributed with master_create_distributed_table but no shards
are created yet for the table (ie master_create_worker_shards not
called), you get unclear error message. This fix catches that case
implicitly and more meaningful message is shown.