Tweak Makefile version detection
As reported [here][1], older versions of make don't support `else if`-
like constructs, so we need to just use plain `if` statements or nest
them if absolutely necessary.

Additionally, we were missing the logic to set `MAJORVERSION` based on
`VERSION` if the former is not present. Older PostgreSQL versions need
this, so I've added it.

[1]: https://groups.google.com/d/msg/pg_shard-users/N92qnf-Cj6A/lqPKw_9ekfUJ