Lock metadata during cluster structural changes
The guiding principle here is that readers must know about a relevant
set of objects to perform distributed queries so need guarantees that
the relevant set will not change beneath them during said queries.

In practice, this means we need two locks: one for the set of shards
related to a relation and one for the set of placements related to a
shard. If the caller intends to add shards (or (active) placements),
they must hold an exclusive lock on the grouping relation (or shard).
If the caller just wants to act on the current set of shards (or the
(active) placements), they need a shared lock before reading.