Use master's port if worker port config absent
CitusDB allows lines in pg_worker_list.conf to omit the port number,
whereas pg_shard required it: confusing (or annoying) users of both.

This changes the sscanf format string to parse both the node name and
node port as strings. If no node port is specified (or if that token
starts with '#', i.e. begins a line comment), the default port is used;
otherwise, we parse the node port string into an integer in precisely
the same fashion used by CitusDB.