Fix hash token distribution algorithm
There are 2^32 distinct "hash tokens" in our hash space, but we were
using INT32_MAX (2^32 - 1) in the code instead. Because of this, shard
counts which one might expect to divide evenly into the space (such as
16, 32, or 256) had fewer tokens than they should have. The remainder
of the tokens were stuffed into the last shard, causing uneven load.

This fixes that by defining a 64-bit constant equal to 2^32 and then
using that to calculate the distance between shards in the hash space.